adapt file-organization to the rest of s5p
