--- drivers/media/platform/s5p-tv/s5p_tvout_common_lib.h
@@ -1,187 +1,0 @@
-/*
- * Copyright (c) 2011 Samsung Electronics Co., Ltd.
- *		http://www.samsung.com
- *
- * Header file of common library for SAMSUNG S5P TVOUT driver
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 2 as
- * published by the Free Software Foundation.
-*/
-
-#ifndef __S5P_TVOUT_COMMON_LIB_H_
-#define __S5P_TVOUT_COMMON_LIB_H_ __FILE__
-
-#include <linux/stddef.h>
-#include <linux/platform_device.h>
-#include <linux/videodev2.h>
-#include <linux/interrupt.h>
-
-/*
- * This file includes declarations for TVOUT driver's common library.
- * All files in TVOUT driver can access function or definition in this file.
- */
-
-#define DRV_NAME	"S5P-TVOUT"
-
-#define tvout_err(fmt, ...)					\
-		printk(KERN_ERR "[%s] %s(): " fmt,		\
-			DRV_NAME, __func__, ##__VA_ARGS__)
-
-#ifndef tvout_dbg
-#ifdef CONFIG_S5P_TVOUT_DEBUG
-#define tvout_dbg(fmt, ...)					\
-		printk(KERN_INFO "[%s] %s(): " fmt,		\
-			DRV_NAME, __func__, ##__VA_ARGS__)
-#else
-#define tvout_dbg(fmt, ...)
-#endif
-#endif
-
-#define S5PTV_FB_CNT		2
-#define HDMI_START_NUM		0x1000
-
-#define to_tvout_plat(d) (to_platform_device(d)->dev.platform_data)
-
-enum s5p_tvout_disp_mode {
-	TVOUT_NTSC_M,
-	TVOUT_PAL_BDGHI,
-	TVOUT_PAL_M,
-	TVOUT_PAL_N,
-	TVOUT_PAL_NC,
-	TVOUT_PAL_60,
-	TVOUT_NTSC_443,
-
-	TVOUT_480P_60_16_9 = HDMI_START_NUM,
-	TVOUT_480P_60_4_3,
-	TVOUT_480P_59,
-
-	TVOUT_576P_50_16_9,
-	TVOUT_576P_50_4_3,
-
-	TVOUT_720P_60,
-	TVOUT_720P_50,
-	TVOUT_720P_59,
-
-	TVOUT_1080P_60,
-	TVOUT_1080P_50,
-	TVOUT_1080P_59,
-	TVOUT_1080P_30,
-
-	TVOUT_1080I_60,
-	TVOUT_1080I_50,
-	TVOUT_1080I_59,
-	TVOUT_INIT_DISP_VALUE
-};
-
-enum s5p_tvout_o_mode {
-	TVOUT_COMPOSITE,
-	TVOUT_HDMI,
-	TVOUT_HDMI_RGB,
-	TVOUT_DVI,
-	TVOUT_INIT_O_VALUE
-};
-
-enum s5p_mixer_burst_mode {
-	MIXER_BURST_8,
-	MIXER_BURST_16,
-};
-
-enum s5ptvfb_data_path_t {
-	DATA_PATH_FIFO,
-	DATA_PATH_DMA,
-};
-
-enum s5ptvfb_alpha_t {
-	LAYER_BLENDING,
-	PIXEL_BLENDING,
-	NONE_BLENDING,
-};
-
-enum s5ptvfb_ver_scaling_t {
-	VERTICAL_X1,
-	VERTICAL_X2,
-};
-
-enum s5ptvfb_hor_scaling_t {
-	HORIZONTAL_X1,
-	HORIZONTAL_X2,
-};
-
-struct s5ptvfb_alpha {
-	enum s5ptvfb_alpha_t	mode;
-	int			channel;
-	unsigned int		value;
-};
-
-struct s5ptvfb_chroma {
-	int		enabled;
-	unsigned int	key;
-};
-
-struct s5ptvfb_user_window {
-	int x;
-	int y;
-};
-
-struct s5ptvfb_user_plane_alpha {
-	int		channel;
-	unsigned char	alpha;
-};
-
-struct s5ptvfb_user_chroma {
-	int		enabled;
-	unsigned char	red;
-	unsigned char	green;
-	unsigned char	blue;
-};
-
-struct s5ptvfb_user_scaling {
-	enum s5ptvfb_ver_scaling_t ver;
-	enum s5ptvfb_hor_scaling_t hor;
-};
-
-struct s5p_tvout_status {
-	struct clk *i2c_phy_clk;
-	struct clk *sclk_hdmiphy;
-	struct clk *sclk_pixel;
-	struct clk *sclk_dac;
-	struct clk *sclk_hdmi;
-};
-
-struct reg_mem_info {
-	char		*name;
-	struct resource *res;
-	void __iomem	*base;
-};
-
-struct irq_info {
-	char		*name;
-	irq_handler_t	handler;
-	int		no;
-};
-
-struct s5p_tvout_clk_info {
-	char		*name;
-	struct clk	*ptr;
-};
-
-extern struct s5p_tvout_status s5ptv_status;
-
-extern int s5p_tvout_vcm_create_unified(void);
-extern int s5p_tvout_vcm_init(void);
-extern void s5p_tvout_vcm_activate(void);
-extern void s5p_tvout_vcm_deactivate(void);
-
-extern int s5p_tvout_map_resource_mem(struct platform_device *pdev,
-				      char *name, void __iomem **base,
-				      struct resource **res);
-extern void s5p_tvout_unmap_resource_mem(void __iomem *base,
-					 struct resource *res);
-
-extern void s5p_tvout_pm_runtime_enable(struct device *dev);
-extern void s5p_tvout_pm_runtime_disable(struct device *dev);
-extern void s5p_tvout_pm_runtime_get(void);
-extern void s5p_tvout_pm_runtime_put(void);
-
-#endif /* __S5P_TVOUT_COMMON_LIB_H_ */
